+++ content/data-structures/HashMap.h
@@ -15,14 +15,14 @@ struct chash { // large odd number for C
 	const uint64_t C = ll(4e18 * acos(0)) | 71;
 	ll operator()(ll x) const { return __builtin_bswap64(x*C); }
 };
-__gnu_pbds::gp_hash_table<ll,int,chash> h({},{},{},{},{1<<16});
+__gnu_pbds::gp_hash_table<ll,ll,chash> h({},{},{},{},{1<<16});
 
 /** For CodeForces, or other places where hacking might be a problem:
 
-const int RANDOM = chrono::high_resolution_clock::now().time_since_epoch().count();
+const ll RANDOM = chrono::high_resolution_clock::now().time_since_epoch().count();
 struct chash { // To use most bits rather than just the lowest ones:
 	const uint64_t C = ll(4e18 * acos(0)) | 71; // large odd number
 	ll operator()(ll x) const { return __builtin_bswap64((x^RANDOM)*C); }
 };
-__gnu_pbds::gp_hash_table<ll, int, chash> h({},{},{},{}, {1 << 16});
+__gnu_pbds::gp_hash_table<ll, ll, chash> h({},{},{},{}, {1 << 16});
 */
