+++ content/data-structures/MoQueries.h
@@ -11,17 +11,17 @@
  */
 #pragma once
 
-void add(int ind, int end) { ... } // add a[ind] (end = 0 or 1)
-void del(int ind, int end) { ... } // remove a[ind]
-int calc() { ... } // compute current answer
+void add(ll ind, ll end) { ... } // add a[ind] (end = 0 or 1)
+void del(ll ind, ll end) { ... } // remove a[ind]
+ll calc() { ... } // compute current answer
 
 vi mo(vector<pii> Q) {
-	int L = 0, R = 0, blk = 350; // ~N/sqrt(Q)
+	ll L = 0, R = 0, blk = 350; // ~N/sqrt(Q)
 	vi s(sz(Q)), res = s;
 #define K(x) pii(x.first/blk, x.second ^ -(x.first/blk & 1))
 	iota(all(s), 0);
-	sort(all(s), [&](int s, int t){ return K(Q[s]) < K(Q[t]); });
-	for (int qi : s) {
+	sort(all(s), [&](ll s, ll t){ return K(Q[s]) < K(Q[t]); });
+	for (ll qi : s) {
 		pii q = Q[qi];
 		while (L > q.first) add(--L, 0);
 		while (R < q.second) add(R++, 1);
@@ -32,24 +32,24 @@ vi mo(vector<pii> Q) {
 	return res;
 }
 
-vi moTree(vector<array<int, 2>> Q, vector<vi>& ed, int root=0){
-	int N = sz(ed), pos[2] = {}, blk = 350; // ~N/sqrt(Q)
+vi moTree(vector<array<ll, 2>> Q, vector<vi>& ed, ll root=0){
+	ll N = sz(ed), pos[2] = {}, blk = 350; // ~N/sqrt(Q)
 	vi s(sz(Q)), res = s, I(N), L(N), R(N), in(N), par(N);
 	add(0, 0), in[0] = 1;
-	auto dfs = [&](int x, int p, int dep, auto& f) -> void {
+	auto dfs = [&](ll x, ll p, ll dep, auto& f) -> void {
 		par[x] = p;
 		L[x] = N;
 		if (dep) I[x] = N++;
-		for (int y : ed[x]) if (y != p) f(y, x, !dep, f);
+		for (ll y : ed[x]) if (y != p) f(y, x, !dep, f);
 		if (!dep) I[x] = N++;
 		R[x] = N;
 	};
 	dfs(root, -1, 0, dfs);
 #define K(x) pii(I[x[0]] / blk, I[x[1]] ^ -(I[x[0]] / blk & 1))
 	iota(all(s), 0);
-	sort(all(s), [&](int s, int t){ return K(Q[s]) < K(Q[t]); });
-	for (int qi : s) rep(end,0,2) {
-		int &a = pos[end], b = Q[qi][end], i = 0;
+	sort(all(s), [&](ll s, ll t){ return K(Q[s]) < K(Q[t]); });
+	for (ll qi : s) rep(end,0,2) {
+		ll &a = pos[end], b = Q[qi][end], i = 0;
 #define step(c) { if (in[c]) { del(a, end); in[a] = 0; } \
                   else { add(c, end); in[c] = 1; } a = c; }
 		while (!(L[b] <= L[a] && R[a] <= R[b]))
