+++ content/data-structures/RMQ.h
@@ -17,15 +17,15 @@ template<class T>
 struct RMQ {
 	vector<vector<T>> jmp;
 	RMQ(const vector<T>& V) : jmp(1, V) {
-		for (int pw = 1, k = 1; pw * 2 <= sz(V); pw *= 2, ++k) {
+		for (ll pw = 1, k = 1; pw * 2 <= sz(V); pw *= 2, ++k) {
 			jmp.emplace_back(sz(V) - pw * 2 + 1);
 			rep(j,0,sz(jmp[k]))
 				jmp[k][j] = min(jmp[k - 1][j], jmp[k - 1][j + pw]);
 		}
 	}
-	T query(int a, int b) {
+	T query(ll a, ll b) {
 		assert(a < b); // or return inf if a == b
-		int dep = 31 - __builtin_clz(b - a);
+		ll dep = 31 - __builtin_clz(b - a);
 		return min(jmp[dep][a], jmp[dep][b - (1 << dep)]);
 	}
 };
