+++ content/geometry/DelaunayTriangulation.h
@@ -15,7 +15,7 @@
 
 template<class P, class F>
 void delaunay(vector<P>& ps, F trifun) {
-	if (sz(ps) == 3) { int d = (ps[0].cross(ps[1], ps[2]) < 0);
+	if (sz(ps) == 3) { ll d = (ps[0].cross(ps[1], ps[2]) < 0);
 		trifun(0,1+d,2-d); }
 	vector<P3> p3;
 	for (P p : ps) p3.emplace_back(p.x, p.y, p.dist2());
