+++ content/geometry/InsidePolygon.h
@@ -20,7 +20,7 @@
 
 template<class P>
 bool inPolygon(vector<P> &p, P a, bool strict = true) {
-	int cnt = 0, n = sz(p);
+	ll cnt = 0, n = sz(p);
 	rep(i,0,n) {
 		P q = p[(i + 1) % n];
 		if (onSegment(p[i], q, a)) return !strict;
