+++ content/geometry/LineHullIntersection.h
@@ -23,13 +23,13 @@
 
 #define cmp(i,j) sgn(dir.perp().cross(poly[(i)%n]-poly[(j)%n]))
 #define extr(i) cmp(i + 1, i) >= 0 && cmp(i, i - 1 + n) < 0
-template <class P> int extrVertex(vector<P>& poly, P dir) {
-	int n = sz(poly), lo = 0, hi = n;
+template <class P> ll extrVertex(vector<P>& poly, P dir) {
+	ll n = sz(poly), lo = 0, hi = n;
 	if (extr(0)) return 0;
 	while (lo + 1 < hi) {
-		int m = (lo + hi) / 2;
+		ll m = (lo + hi) / 2;
 		if (extr(m)) return m;
-		int ls = cmp(lo + 1, lo), ms = cmp(m + 1, m);
+		ll ls = cmp(lo + 1, lo), ms = cmp(m + 1, m);
 		(ls < ms || (ls == ms && ls == cmp(lo, m)) ? hi : lo) = m;
 	}
 	return lo;
@@ -37,16 +37,16 @@ template <class P> int extrVertex(vector<P>& poly, P dir) {
 
 #define cmpL(i) sgn(a.cross(poly[i], b))
 template <class P>
-array<int, 2> lineHull(P a, P b, vector<P>& poly) {
-	int endA = extrVertex(poly, (a - b).perp());
-	int endB = extrVertex(poly, (b - a).perp());
+array<ll, 2> lineHull(P a, P b, vector<P>& poly) {
+	ll endA = extrVertex(poly, (a - b).perp());
+	ll endB = extrVertex(poly, (b - a).perp());
 	if (cmpL(endA) < 0 || cmpL(endB) > 0)
 		return {-1, -1};
-	array<int, 2> res;
+	array<ll, 2> res;
 	rep(i,0,2) {
-		int lo = endB, hi = endA, n = sz(poly);
+		ll lo = endB, hi = endA, n = sz(poly);
 		while ((lo + 1) % n != hi) {
-			int m = ((lo + hi + (lo < hi ? 0 : n)) / 2) % n;
+			ll m = ((lo + hi + (lo < hi ? 0 : n)) / 2) % n;
 			(cmpL(m) == cmpL(endB) ? lo : hi) = m;
 		}
 		res[i] = (lo + !cmpL(hi)) % n;
