+++ content/geometry/Point.h
@@ -4,12 +4,12 @@
  * License: CC0
  * Source: My head with inspiration from tinyKACTL
  * Description: Class to handle points in the plane.
- * 	T can be e.g. double or long long. (Avoid int.)
+ * 	T can be e.g. double or long long. (Avoid ll.)
  * Status: Works fine, used a lot
  */
 #pragma once
 
-template <class T> int sgn(T x) { return (x > 0) - (x < 0); }
+template <class T> ll sgn(T x) { return (x > 0) - (x < 0); }
 template<class T>
 struct Point {
 	typedef Point P;
