+++ content/geometry/PointInsideHull.h
@@ -19,13 +19,13 @@
 typedef Point<ll> P;
 
 bool inHull(const vector<P>& l, P p, bool strict = true) {
-	int a = 1, b = sz(l) - 1, r = !strict;
+	ll a = 1, b = sz(l) - 1, r = !strict;
 	if (sz(l) < 3) return r && onSegment(l[0], l.back(), p);
 	if (sideOf(l[0], l[a], l[b]) > 0) swap(a, b);
 	if (sideOf(l[0], l[a], p) >= r || sideOf(l[0], l[b], p)<= -r)
 		return false;
 	while (abs(a - b) > 1) {
-		int c = (a + b) / 2;
+		ll c = (a + b) / 2;
 		(sideOf(l[0], l[c], p) > 0 ? b : a) = c;
 	}
 	return sgn(l[a].cross(l[b], p)) < r;
