+++ content/geometry/PolygonCenter.h
@@ -14,7 +14,7 @@
 typedef Point<double> P;
 P polygonCenter(const vector<P>& v) {
 	P res(0, 0); double A = 0;
-	for (int i = 0, j = sz(v) - 1; i < sz(v); j = i++) {
+	for (ll i = 0, j = sz(v) - 1; i < sz(v); j = i++) {
 		res = res + (v[i] + v[j]) * v[j].cross(v[i]);
 		A += v[j].cross(v[i]);
 	}
