+++ content/geometry/PolygonUnion.h
@@ -20,11 +20,11 @@ double polyUnion(vector<vector<P>>& poly) {
 	double ret = 0;
 	rep(i,0,sz(poly)) rep(v,0,sz(poly[i])) {
 		P A = poly[i][v], B = poly[i][(v + 1) % sz(poly[i])];
-		vector<pair<double, int>> segs = {{0, 0}, {1, 0}};
+		vector<pair<double, ll>> segs = {{0, 0}, {1, 0}};
 		rep(j,0,sz(poly)) if (i != j) {
 			rep(u,0,sz(poly[j])) {
 				P C = poly[j][u], D = poly[j][(u + 1) % sz(poly[j])];
-				int sc = sideOf(A, B, C), sd = sideOf(A, B, D);
+				ll sc = sideOf(A, B, C), sd = sideOf(A, B, D);
 				if (sc != sd) {
 					double sa = C.cross(D, A), sb = C.cross(D, B);
 					if (min(sc, sd) < 0)
@@ -38,7 +38,7 @@ double polyUnion(vector<vector<P>>& poly) {
 		sort(all(segs));
 		for (auto& s : segs) s.first = min(max(s.first, 0.0), 1.0);
 		double sum = 0;
-		int cnt = segs[0].second;
+		ll cnt = segs[0].second;
 		rep(j,1,sz(segs)) {
 			if (!cnt) sum += segs[j].first - segs[j - 1].first;
 			cnt += segs[j].second;
